+++ models/vote.go
@@ -2,6 +2,6 @@ package models
 
 //easyjson:json
 type Vote struct {
-	Nickname string  `json:"nickname"`
-	Voice    float32 `json:"voice"`
+	Nickname string `json:"nickname"`
+	Voice    int    `json:"voice"`
 }
